% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{dimensionar}
\alias{dimensionar}
\title{Determinar las dimensiones de una matriz a partir de una cantidad de elementos}
\usage{
dimensionar(l)
}
\arguments{
\item{l}{Cantidad total de elementos.}
}
\value{
Un vector numérico de longitud 2 con el número de filas y columnas.
}
\description{
Calcula dimensiones preferentemente cuadradas para una matriz que contenga
una cantidad determinada de elementos. Si no es posible obtener una matriz
cuadrada exacta, se prefiere una forma más ancha que alta.
}
\keyword{internal}
