% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMemoryFeatures.R
\name{extractMemoryFeatures}
\alias{extractMemoryFeatures}
\title{Extracts ecological memory features on the output of \code{\link{computeMemory}}.}
\usage{
extractMemoryFeatures(
 analysis.output = NULL,
 exogenous.component = NULL,
 endogenous.component = NULL,
 sampling.subset = NULL
 )
}
\arguments{
\item{analysis.output}{dataframe, output of \code{\link{computeMemory}}.}

\item{exogenous.component}{character string, name of the variable defining the exogenous component.}

\item{endogenous.component}{character string, string, name of the variable defining the endogenous component.}

\item{sampling.subset}{only used when analysis.output is the result of runExperiment(). String with the dataset type, one of: "Annual", "1cm", "2cm", "6cm", "10cm".}
}
\value{
A list with three 4 slots:
}
\description{
It computes the following features of the ecological memory patterns returned by \code{\link{computeMemory}}:
\itemize{
   \item \code{memory strength} maximum difference in relative importance between each component (endogenous, exogenous, and concurrent) and the median of the random component. This is computed for exogenous, endogenous, and concurrent effect.
    \item \code{memory length} proportion of lags over which the importance of a memory component is above the median of the random component. This is only computed for endogenous and exogenous memory.
\item \code{dominance} proportion of the lags above the median of the random term over which a memory component has a higher importance than the other component. This is only computed for endogenous and exogenous memory.
}
}
\details{

}
\examples{

}
\seealso{
\code{\link{computeMemory}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
