% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoria.R
\docType{data}
\name{laggedSimData}
\alias{laggedSimData}
\title{Lagged data generated by \code{\link{prepareLaggedData}}}
\format{dataframe with 203 columns and 9900 rows.}
\usage{
data(laggedSimData)
}
\description{
A dataframe resulting from the application of \code{\link{prepareLaggedData}} to the first slot of the list \code{\link{simulation}}. The order of the cases represent age. The first row contains the "older" sample, while the last row contains the "newest one". Therefore, for a lag equal to 1, any given case of the column \code{Response_0} is aligned with the data of the row above it. The dataframe columns are:
}
\details{
\itemize{
  \item \emph{Response_0} numeric, values of the Response (simulated pollen abundance in this case) for the lag 0 (meaning that data in this column has not been lagged). This column is used as response variable by the function \code{\link{computeMemory}}.
  \item \emph{Response_1-100} numeric, time delayed values of the response for different lags (in years). Considered together these columns represent the endogenous ecological memory.
  \item \emph{Driver.A_0} numeric, values of the variable Driver A for the lag 0 (no lag). This column represents the concurrent effect of the driver over the response.
  \item \emph{Driver.A_1-100} numeric, time delayed values of Driver.A for different lags (in years). Considered together these columns represent the exogenous ecological memory.
}
}
\seealso{
\code{\link{simulatePopulation}}, \code{\link{plotSimulation}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
\keyword{datasets}
