% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{limpiar_consola}
\alias{limpiar_consola}
\title{Limpiar la consola}
\usage{
limpiar_consola()
}
\value{
No devuelve valor.
}
\description{
Limpia el contenido de la consola, compatible con sistemas Windows y Unix.
}
\examples{
limpiar_consola()

}
