% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{mostrar_info}
\alias{mostrar_info}
\title{Mostrar información del estado actual de la partida}
\usage{
mostrar_info(tablero, afuera, ronda)
}
\arguments{
\item{tablero}{Matriz de fichas generada por \code{nuevo_tablero()}.}

\item{afuera}{Vector de posiciones (enteros) que ya fueron descubiertas.}

\item{ronda}{Número de ronda actual.}
}
\value{
No devuelve valor. Imprime información en consola.
}
\description{
Imprime información útil sobre la partida actual, incluyendo la ronda, pares formados y pares restantes.
}
\examples{
fichas <- c("a", "b", "c")
tablero <- nuevo_tablero(fichas, semilla = 1)
mostrar_tablero(tablero)
mostrar_tablero(tablero, mostrar = c(2, 6))
mostrar_info(tablero, afuera = c(2, 6), ronda = 1)

}
