% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{mostrar_pares}
\alias{mostrar_pares}
\title{Generar texto con pares de fichas}
\usage{
mostrar_pares(fichas = character())
}
\arguments{
\item{fichas}{Vector de caracteres con fichas únicas.}
}
\value{
Vector de texto con pares formateados.
}
\description{
Crea un vector de cadenas con la forma (ficha, ficha) para cada ficha del vector.
}
\examples{
fichas <- c("a", "b", "c")
mostrar_pares(fichas)

}
