% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{mostrar_tablero}
\alias{mostrar_tablero}
\title{Mostrar el tablero con fichas ocultas o reveladas}
\usage{
mostrar_tablero(tablero, mostrar = NULL)
}
\arguments{
\item{tablero}{Matriz generada por \code{nuevo_tablero()}.}

\item{mostrar}{Vector numérico con las posiciones de las fichas a mostrar. Estas posiciones se corresponden con los números de fichas que se ven al mostrar el tablero en la consola. No se requiere indexar fila y columna, sólo proveer el número de ficha.}
}
\value{
No devuelve valor. Imprime el tablero formateado en consola.
}
\description{
Imprime el tablero en la consola, ocultando las fichas no seleccionadas y mostrando aquellas indicadas.
}
\details{
Esta función limpia la consola antes de imprimir.
}
\examples{
fichas <- c("a", "b", "c")
tablero <- nuevo_tablero(fichas)
mostrar_tablero(tablero)
mostrar_tablero(tablero, mostrar = c(2, 6))
}
