% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{nuevo_tablero}
\alias{nuevo_tablero}
\title{Crear un nuevo tablero con pares de fichas mezcladas aleatoriamente}
\usage{
nuevo_tablero(fichas, semilla = NULL)
}
\arguments{
\item{fichas}{Vector de caracteres con las fichas únicas.}

\item{semilla}{Entero opcional para fijar la semilla de aleatorización, para
que el tablero luzca siempre igual (no debe ser usado en la versión final).}
}
\value{
Una matriz con las fichas dispuestas en posiciones aleatorias. Las
  dimensiones se eligen automáticamente.
}
\description{
Genera una matriz con pares de fichas dispuestas en orden aleatorio. Las
fichas se duplican y se organizan en forma matricial.
}
\examples{
fichas <- c("a", "b", "c")
tablero <- nuevo_tablero(fichas)
tablero

}
