% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoria.R
\docType{data}
\name{palaeodata}
\alias{palaeodata}
\title{Dataframe with pollen and climate data.}
\format{dataframe with 10 columns and 7986 rows.}
\usage{
data(palaeodata)
}
\description{
A dataframe with a regular time grid of 0.2 ky resolution resulting from applying \code{\link{mergePalaeoData}} to the datasets \code{\link{climate}} and \code{\link{pollen}}:
}
\details{
\itemize{
  \item \emph{age} in ky before present (ky BP).
  \item \emph{pinus} pollen counts of Pinus.
  \item \emph{quercus} pollen counts of Quercus.
  \item \emph{poaceae} pollen counts of Poaceae.
  \item \emph{artemisia} pollen counts of Artemisia.
  \item \emph{temperatureAverage} average annual temperature in Celsius degrees.
  \item \emph{rainfallAverage} average annual precipitation in milimetres per day (mm/day).
  \item \emph{temperatureWarmestMonth} average temperature of the warmest month, in Celsius degrees.
  \item \emph{temperatureColdestMonth} average temperature of the coldest month, in Celsius degrees.
  \item \emph{oxigenIsotope} delta O18, global ratio of stable isotopes in the sea floor, see \url{http://lorraine-lisiecki.com/stack.html} for further details.
}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
\keyword{datasets}
