% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExperiment.R
\name{plotExperiment}
\alias{plotExperiment}
\title{Plots the output of  \code{\link{runExperiment}}.}
\usage{
plotExperiment(
  experiment.output = NULL,
  parameters.file = NULL,
  ribbon = FALSE
)
}
\arguments{
\item{experiment.output}{list, output of  \code{\link{runExperiment}}. Default: \code{NULL}.}

\item{parameters.file}{dataframe of simulation parameters. Default: \code{NULL}.}

\item{ribbon}{logical, switches plotting of confidence intervals on (TRUE) and off (FALSE). Default: \code{FALSE}.}
}
\value{
A ggplot2 object.
}
\description{
Takes the output of \code{\link{runExperiment}}, and generates plots of ecological memory patterns for a large number of simulated pollen curves.
}
\seealso{
\code{\link{plotMemory}}, \code{\link{runExperiment}}

Other virtualPollen: 
\code{\link{experimentToTable}()},
\code{\link{runExperiment}()}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
\concept{virtualPollen}
