% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoria.R
\docType{data}
\name{pollen}
\alias{pollen}
\title{Dataframe with pollen counts.}
\format{
dataframe with 5 columns and 639 rows.
}
\usage{
data(pollen)
}
\description{
A dataframe with the following columns:
}
\details{
\itemize{
  \item \emph{age} in kiloyears before present (ky BP).
  \item \emph{pinus} pollen counts of Pinus.
  \item \emph{quercus} pollen counts of Quercus.
  \item \emph{poaceae} pollen counts of Poaceae.
  \item \emph{artemisia} pollen counts of Artemisia.
}
}
\seealso{
Other example_data: 
\code{\link{climate}},
\code{\link{palaeodata}},
\code{\link{palaeodataLagged}},
\code{\link{palaeodataMemory}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
\concept{example_data}
\keyword{datasets}
