% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{posiciones_disponibles}
\alias{posiciones_disponibles}
\title{Obtener posiciones disponibles en el tablero}
\usage{
posiciones_disponibles(tablero, afuera)
}
\arguments{
\item{tablero}{Tablero de fichas.}

\item{afuera}{Vector numérico de posiciones que ya han sido descubiertas.}
}
\value{
Vector numérico con las posiciones aún disponibles.
}
\description{
Devuelve las posiciones que aún no han sido descubiertas.
}
\examples{
fichas <- c("a", "b", "c")
tablero <- nuevo_tablero(fichas, semilla = 1)
mostrar_tablero(tablero, mostrar = c(2, 6))
posiciones_disponibles(tablero, afuera = c(2, 6))

}
