% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareLaggedData.R
\name{prepareLaggedData}
\alias{prepareLaggedData}
\title{Organizes time series data into lags.}
\usage{
prepareLaggedData(
  input.data = NULL,
  response = NULL,
  drivers = NULL,
  time = NULL,
  lags = seq(0, 200, by=20),
  time.zoom=NULL,
  scale=FALSE
  )
}
\arguments{
\item{input.data}{a dataframe with one time series per column.}

\item{response}{character string, name of the numeric column to be used as response in the model.}

\item{drivers}{character vector, names of the numeric columns to be used as predictors in the model.}

\item{time}{character vector, name of the numeric column with the time/age.}

\item{lags}{numeric vector, lags to be used in the equation. Generally, a regular sequence of numbers. The use \code{\link{seq}} to define it is highly recommended. If 0 is absent from lags, it is added automatically to allow the consideration of a concurrent effect.}

\item{time.zoom}{numeric vector of two numbers of the \code{time} column used to subset the data if desired.}

\item{scale}{boolean, if TRUE, applies the \code{\link{scale}} function to normalize the data. Required if the lagged data is going to be used to fit linear models.}
}
\value{
A dataframe with columns representing time-delayed values of the drivers and the response.
}
\description{
Takes a multivariate time series, where at least one variable is meant to be used as a response while the others are meant to be used as predictors in a model, and organizes it in to quantify ecological memory through models of the form:
 \eqn{p_{t} ~ p_{t-1} +...+ p_{t-n} + d_{t} + d_{t-1} +...+ d_{t-n}}

 where:

\itemize{
 \item \eqn{d} is a driver (several drivers can be added).
 \item \eqn{t} is the time of any given value of the response \emph{p}.
 \item \eqn{t-1} is the lag number 1 (in time units).
 \item \eqn{p_{t-1} +...+ p_{t-n}}  represents the endogenous component of ecological memory.
 \item \eqn{d_{t-1} +...+ d_{t-n}}  represents the exogenous component of ecological memory.
 \item \eqn{d_{t}} represents the concurrent effect of the driver over the response.
}
}
\details{
The function returns a dataframe. Column names have the lag number as a suffix. The response variable is identified by changing its name to "Response".
}
\examples{
#loading data
data(simulation)

#adding lags
sim.lags <- prepareLaggedData(
  input.data = simulation[[1]],
  response = "Pollen",
  drivers = "Driver",
  time = "Time",
  lags = seq(0, 200, by=20),
  time.zoom=NULL,
  scale=FALSE
  )

}
\seealso{
\code{\link{computeMemory}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
