% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{str_timer}
\alias{str_timer}
\title{Mostrar un texto car치cter por car치cter con retardo}
\usage{
str_timer(s, t = 0.05)
}
\arguments{
\item{s}{Cadena de texto a imprimir.}

\item{t}{Tiempo de espera entre caracteres (en segundos).}
}
\value{
No devuelve valor. Efecto visual en consola.
}
\description{
Imprime un texto en la consola, revelando cada car치cter con un peque침o
retardo.
}
\keyword{internal}
