% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{texto_lento}
\alias{texto_lento}
\title{Imprimir texto lentamente en la consola}
\usage{
texto_lento(texto)
}
\arguments{
\item{texto}{Vector o lista de cadenas de texto a imprimir. Puede ser una
sola cadena de texto.}
}
\value{
No devuelve valor. Efecto visual en consola.
}
\description{
Imprime cada línea de texto de forma pausada, carácter por carácter.
}
\examples{
texto_lento("Hola... ¿Estás listo para jugar?")

}
