% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{validar_fichas}
\alias{validar_fichas}
\title{Validar un conjunto de fichas ingresadas}
\usage{
validar_fichas(fichas, n)
}
\arguments{
\item{fichas}{Vector de caracteres con las fichas ingresadas.}

\item{n}{Número mínimo de fichas requerido.}
}
\value{
TRUE si las fichas son válidas, FALSE en caso contrario.
}
\description{
Verifica que la cantidad de fichas sea suficiente y que no haya fichas duplicadas.
}
\examples{
fichas <- c("a", "b", "c")
validar_fichas(fichas, 4)
validar_fichas(fichas, 3)
fichas <- c("a", "b", "c", "c")
validar_fichas(fichas, 4)

}
