% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentToTable.R
\name{experimentToTable}
\alias{experimentToTable}
\title{Turns the outcome of \code{\link{runExperiment}} into a long table.}
\usage{
experimentToTable(
 experiment.output = NULL,
 parameters.file = NULL,
 sampling.names = NULL,
 R2 = TRUE
 )
}
\arguments{
\item{experiment.output}{list, output of  \code{\link{runExperiment}}.}

\item{parameters.file}{dataframe of simulation parameters for \code{\link[virtualPollen]{simulatePopulation}}. Generally produced by \code{\link[virtualPollen]{parametersDataframe}}.}

\item{sampling.names}{vector of character strings with the names of the columns of \code{simulations.file}.}

\item{R2}{boolean. If TRUE, the average pseudo R-squared of the random forest models used to analyze the ecological memory pattern of the virtual taxa are shown with the taxon traits.}
}
\value{
A dataframe.
}
\description{
Takes the output of \code{\link{runExperiment}}, extracts the dataframes containing the ecological memory patterns generated by \code{\link{computeMemory}}, and binds them together into a single dataframe ready for further analyses or plotting.
}
\details{
This function is used internally by \code{\link{plotExperiment}}, but it is also available to users in case they want to do other kinds of analyses or plots with the data.
}
\examples{

}
\seealso{
\code{\link{runExperiment}}, \code{\link{plotExperiment}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
