% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoria.R
\docType{data}
\name{palaeodata}
\alias{palaeodata}
\title{Dataframe with pollen and climate data.}
\format{
dataframe with 10 columns and 7986 rows.
}
\usage{
data(palaeodata)
}
\description{
A dataframe with a regular time grid of 0.2 ky resolution resulting from applying \code{\link{mergePalaeoData}} to the datasets \code{\link{climate}} and \code{\link{pollen}}:
}
\details{
\itemize{
  \item \emph{age} in ky before present (ky BP).
  \item \emph{pollen.pinus} pollen percentages of Pinus.
  \item \emph{pollen.quercus} pollen percentages of Quercus.
  \item \emph{pollen.poaceae} pollen percentages of Poaceae.
  \item \emph{pollen.artemisia} pollen percentages of Artemisia.
  \item \emph{climate.temperatureAverage} average annual temperature in degrees Celsius.
  \item \emph{climate.rainfallAverage} average annual precipitation in millimetres per day (mm/day).
  \item \emph{climate.temperatureWarmestMonth} average temperature of the warmest month, in degrees Celsius.
  \item \emph{climate.temperatureColdestMonth} average temperature of the coldest month, in degrees Celsius.
  \item \emph{climate.oxigenIsotope} delta O18, global ratio of stable isotopes in the sea floor, see \url{http://lorraine-lisiecki.com/stack.html} for further details.
}
}
\seealso{
Other example_data: 
\code{\link{climate}},
\code{\link{palaeodataLagged}},
\code{\link{palaeodataMemory}},
\code{\link{pollen}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
\concept{example_data}
\keyword{datasets}
