% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoria.R
\docType{data}
\name{pollen}
\alias{pollen}
\title{Dataframe with pollen counts.}
\format{dataframe with 10 columns and 7986 rows.}
\usage{
data(pollen)
}
\description{
A dataframe with the following columns:
}
\details{
\itemize{
  \item \emph{age} in kiloyears before present (ky BP).
  \item \emph{pinus} pollen counts of Pinus.
  \item \emph{quercus} pollen counts of Quercus.
  \item \emph{poaceae} pollen counts of Poaceae.
  \item \emph{artemisia} pollen counts of Artemisia.
}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
\keyword{datasets}
